+++ src/diagnostics/tsx/validators/validator.ts
@@ -1,0 +1,7 @@
+import { Diagnostic } from '../diagnostic';
+import { TSXElement } from '../element';
+
+export interface Validator {
+  readonly tags: readonly string[];
+  validate(node: TSXElement): Diagnostic[];
+}